Fix ca cert removal

+++ site-modules/bolt/tasks/remove_ca_cert.sh
@@ -1,20 +1,24 @@
 #!/bin/bash
 
 # check if we are running on a PE server
-# which rpm
+hostnamectl | grep 'Operating System: Ubuntu'
 
-# if [ $? -eq 0 ]
-# then
-#   rpm -qa | grep pe-puppetserver
-# else
-#   dpkg -l | grep pe-puppetserver
-# fi
+if [ $? -eq 0 ]
+then
+  echo 'ubuntu'
+  command='dpkg -l'
+else
+  echo 'redhat'
+  command='rpm -qa'
+fi
 
-# if [ $? -ne 0 ]
-# then
-#   echo "Bailing out: Running on PE server"
-#   exit 1
-# fi
+$command | grep pe-puppetserver
+
+if [ $? -eq 0 ]
+then
+  echo "Bailing out: Running on PE server"
+  exit 1
+fi
 
 # Remove ca cert
 rm /etc/puppetlabs/puppet/ssl/certs/ca.pem
